Pre: Move core/Attribute* to core-traits
