+++ query-algebrizer/src/clauses/or.rs
@@ -754,12 +754,12 @@ mod testing {
     use super::*;
 
     use core_traits::{
+        Attribute,
         ValueType,
         TypedValue,
     };
 
     use mentat_core::{
-        Attribute,
         Schema,
     };
